refactor(route-service): Renamed services
Renamed servises to kebab-case
Added to services `.service` suffix

Removed ununsed full-commented file
